back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langgraph>=1.0.7",
    "openai>=2.16.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.2",
    "uvicorn>=0.40.0",
    "websockets>=11,<16",
]
